New version of 'dalton' data with 'tritan' support.
Ken contacted Francoise to provide this in the format of Table 2 in the
CR&A99 paper (that only lists deutan/protan). She provided a table with
RGB values _not_ rounded to integers. Hence, the RGB values for deutan/protan
are slightly different (but identical when rounded) compared to earlier
versions of dalton. The tritan values are new in this version.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/colorspace@96 edb9625f-4e0d-4859-8d74-9fd3b1da38cb